Rewrite descriptors with atomics
The memory ordering ensures that operations to normal memory are
synchronized with operations on device memory, at runtime. I've seen
this play out in the transmit path: writes to the transmit buffer's
flags weren't reaching memory by the time the MAC was checking,
resulting in missing packets. Moving the fence works, but it's better to
use atomics.